improve clients naming and documentation
