+++ back/src/clients/jacket-client/models/jcategory.model.ts
