+++ back/src/clients/jacket-client/models/jsearch.model.ts
