+++ back/src/clients/jacket-client/models/jtorrent.model.ts
