+++ back/src/clients/transmission-daemon-client/models/tdconfig.model.ts
@@ -1,0 +1,5 @@
+export interface TDConfig {
+  url: string;
+  login: string;
+  password: string;
+}
