+++ back/src/clients/transmission-daemon-client/models/tdrequest.model.ts
@@ -1,0 +1,4 @@
+export interface TDRequest {
+  arguments: any;
+  method: string;
+}
